Fix recognition date from membership renewal form (master only regression)
This was getting set deep in the BAO based on membership_id being passed in - but we dropped that
so need to make sure the form passes it in

+++ tests/phpunit/CRM/Event/Form/ParticipantTest.php
@@ -712,7 +712,7 @@ London,',
    */
   public function testSubmitWithDeferredRecognition(): void {
     Civi::settings()->set('deferred_revenue_enabled', TRUE);
-    $futureDate = date('Y') + 1 . '-09-20';
+    $futureDate = ((int) date('Y')) + 1 . '-09-20';
     $this->submitForm(['is_monetary' => 1, 'financial_type_id' => 1, 'start_date' => $futureDate], [
       'record_contribution' => TRUE,
       'financial_type_id' => 1,
